refactor LocalVideoSelection
