+++ wallpaper-play/Page/Setting/Contents/LocalVideo/LocalVideoSelectionUseCase.swift
@@ -1,0 +1,27 @@
+import Foundation
+import Injectable
+
+struct VideoConfigInput {
+    let link: URL
+    let mute: Bool
+    let videoSize: VideoSize
+}
+
+typealias VideoConfigOutput = VideoConfigInput
+
+protocol LocalVideoSelectionUseCase {
+    func requestSettingWallpaper(_ input: VideoConfigInput)
+}
+
+class LocalVideoSelectionInteractor: LocalVideoSelectionUseCase {
+    
+    private let wallpaperRequestService: WallpaperRequestService
+
+    init(wallpaperRequestService: WallpaperRequestService) {
+        self.wallpaperRequestService = wallpaperRequestService
+    }
+
+    func requestSettingWallpaper(_ input: VideoConfigInput) {
+        wallpaperRequestService.requestVideoWallpaper(video: VideoPlayValue(url: input.link, mute: input.mute, videoSize: input.videoSize))
+    }
+}
